add: base session controller
